Preview is set up, doesn't work.
This passes key and columns data to preview.html via URL parameters, but it fails to load the preview with an Uncaught TypeError: Cannot read property 'length' of undefined.

I've tested it with jQuery 1.11 unminimized to try to figure out the problem, but it fails with another TypeError: Uncaught TypeError: Cannot use 'in' operator to search for '98' in [["word","Word"],["use","Use"],["first-use","First Use"],["meaning","Meaning"],["origin","Origin"]] in code.jquery.com/jquery-1.11.0.js:592.